requirements.txt
angles==1.9.12
asgiref==3.4.1
bondpy==1.8.5
camera_calibration==1.15.0
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2021.10.8
charset-normalizer==2.0.7
controller_manager==0.18.3
controller_manager_msgs==0.18.3
coreapi==2.3.3
coreschema==0.0.4
cv_bridge==1.13.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
dynamic_reconfigure==1.6.3
Faker==9.8.2
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
gunicorn==20.1.0
idna==3.3
image_geometry==1.13.0
imagesift==1.2.15
inflection==0.5.1
interactive-markers==1.11.5
itypes==1.2.0
Jinja2==3.0.3
joint_state_publisher==1.12.15
joy_listener==0.2.6
joy_mouse==0.1.15
jsk-perception==1.2.15
jsk-recognition-utils==1.2.15
jsk_data==2.2.11
jsk_topic_tools==2.2.11
kdl_parser_py==1.13.1
laser_geometry==1.6.7
MarkupSafe==2.0.1
message_filters==1.14.11
openni2_launch==1.5.1
packaging==21.2
psycopg2==2.9.2
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python_qt_binding==0.4.3
pytz==2021.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
resource_retriever==1.12.6
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.13
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.2
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_rviz==0.6.0
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
rviz==1.13.17
sensor-msgs==1.12.8
six==1.16.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
sound-play==0.3.11
sqlparse==0.4.2
srdfdom==0.5.2
text-unidecode==1.3
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.11
urdfdom-py==0.4.5
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
wiimote==1.14.0
xacro==1.13.10
